+++ tools/google_auth.py
@@ -29,6 +29,7 @@ SCOPES = [
     "https://www.googleapis.com/auth/drive.readonly",
     "https://www.googleapis.com/auth/contacts.readonly",
     "https://www.googleapis.com/auth/contacts.other.readonly",
+    "https://www.googleapis.com/auth/documents",
 ]